Landing page and sign up forms
I have created a landing page, for the site. Still planning on adding more information to the page.
I have also setup  a working sign up form for the website to allow users sign up, sign in and/or recover their account. Of course it is just a design with minimal javascript for basic validation.